requirements.txt
colour
edge-tts
Pillow
playsound
pyaudio
PyObjC
setuptools
SpeechRecognition
